# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      operationId: getMyProfile
      summary: Get the current student's profile
      responses:
        "200":
          description: Profile resolved (may be null before onboarding).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfileEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [Profile]
      operationId: upsertMyProfile
      summary: Create or update the current student's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertStudentProfileBody"
      responses:
        "200":
          description: Profile saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CourseType:
      type: string
      enum: [literature_atar, english_atar, general_english]
...
    StudentProfile:
      type: object
      required: [userId, yearLevel, courseType, school, updatedAt]
      properties:
        userId:
          type: string
        yearLevel:
          type: integer
          minimum: 7
          maximum: 12
        courseType:
          $ref: "#/components/schemas/CourseType"
        school:
          type: [string, "null"]
        updatedAt:
          type: string
          format: date-time
    StudentProfileEnvelope:
      type: object
      required: [profile]
      properties:
        profile:
          oneOf:
            - $ref: "#/components/schemas/StudentProfile"
            - type: "null"
    UpsertStudentProfileBody:
      type: object
      required: [yearLevel, courseType]
      properties:
        yearLevel:
          type: integer
          minimum: 7
          maximum: 12
        courseType:
          $ref: "#/components/schemas/CourseType"
        school:
          type: [string, "null"]
...